Update MovieCard snapshot test
